Fix unittests and missed some spots cleaning up
