+++ Tests/PacketHandling_Test.cpp
@@ -10,27 +10,12 @@
 #include "../Includes/PCapReader.h"
 #include "../Includes/XLinkKaiConnection.h"
 #include "IConnectorMock.h"
+#include "IPCapDeviceMock.h"
 
 using ::testing::_;
 using ::testing::DoAll;
 using ::testing::Return;
 using ::testing::WithArg;
-
-class IPCapDeviceMock : public IPCapDevice
-{
-public:
-    MOCK_METHOD(void, BlackList, (uint64_t aMac));
-    MOCK_METHOD(void, Close, ());
-    MOCK_METHOD(bool, Connect, (std::string_view aESSID));
-    MOCK_METHOD(bool, Open, (std::string_view aName, std::vector<std::string>& aSSIDFilter));
-    MOCK_METHOD(std::string, DataToString, (const unsigned char* aData, const pcap_pkthdr* aHeader));
-    MOCK_METHOD(const unsigned char*, GetData, ());
-    MOCK_METHOD(const pcap_pkthdr*, GetHeader, ());
-    MOCK_METHOD(bool, Send, (std::string_view aData));
-    MOCK_METHOD(void, SetConnector, (std::shared_ptr<IConnector> aDevice));
-    MOCK_METHOD(void, SetHosting, (bool aHosting));
-    MOCK_METHOD(bool, StartReceiverThread, ());
-};
 
 class PCapReaderDerived : public PCapReader
 {
